requirements.dev.txt
boto3-stubs[macie,macie2]
